% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keynote.R
\docType{package}
\name{keynote}
\alias{keynote}
\title{Create and Update Keynote Presentations from R}
\format{An object of class \code{keynote} (inherits from \code{R6}) of length 11.}
\usage{
keynote
}
\description{
Create and Update Keynote Presentations from R

TODO
}
\keyword{datasets}
