% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{keynote_document}
\alias{keynote_document}
\title{A Keynote document}
\description{
TODO
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{keynote_document$new()}}
\item \href{#method-run_script}{\code{keynote_document$run_script()}}
\item \href{#method-save}{\code{keynote_document$save()}}
\item \href{#method-properties}{\code{keynote_document$properties()}}
\item \href{#method-filename}{\code{keynote_document$filename()}}
\item \href{#method-close}{\code{keynote_document$close()}}
\item \href{#method-num_slides}{\code{keynote_document$num_slides()}}
\item \href{#method-is_open}{\code{keynote_document$is_open()}}
\item \href{#method-export}{\code{keynote_document$export()}}
\item \href{#method-list_master_slides}{\code{keynote_document$list_master_slides()}}
\item \href{#method-make_slide}{\code{keynote_document$make_slide()}}
\item \href{#method-slide_summary}{\code{keynote_document$slide_summary()}}
\item \href{#method-get_slide}{\code{keynote_document$get_slide()}}
\item \href{#method-clone}{\code{keynote_document$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$new(filename = NULL, open = TRUE, theme = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_script"></a>}}
\subsection{Method \code{run_script()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$run_script(
  text,
  args = character(),
  output_format = c("json", "raw")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$save(filename = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-properties"></a>}}
\subsection{Method \code{properties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$properties()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filename"></a>}}
\subsection{Method \code{filename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$filename()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-num_slides"></a>}}
\subsection{Method \code{num_slides()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$num_slides()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_open"></a>}}
\subsection{Method \code{is_open()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$is_open()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export"></a>}}
\subsection{Method \code{export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$export()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_master_slides"></a>}}
\subsection{Method \code{list_master_slides()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$list_master_slides()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_slide"></a>}}
\subsection{Method \code{make_slide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$make_slide()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-slide_summary"></a>}}
\subsection{Method \code{slide_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$slide_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_slide"></a>}}
\subsection{Method \code{get_slide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$get_slide(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{keynote_document$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
