% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{new_keynote}
\alias{new_keynote}
\title{Create a new Keynote document}
\usage{
new_keynote(theme = NULL)
}
\arguments{
\item{theme}{Theme to use. See [`keynote$list_themes()`][keynote_document]
for the available themes.}
}
\value{
`keynote` object.
}
\description{
Create a new Keynote document
}
