% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{open_keynote}
\alias{open_keynote}
\title{Open a Keynote document}
\usage{
open_keynote(filename = NULL)
}
\arguments{
\item{filename}{path to the document to open.}
}
\value{
`keynote` object.
}
\description{
Open a Keynote document
}
